pyproject.toml
[project]
name = "medienrechtspraxis"
version = "0.1.0"
description = "AI-generated Swiss media law commentary pipeline"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
    "click>=8.1.0",
    "rich>=13.9.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
